requirements.txt
django
django-mptt
django-mptt-admin
django-import-export